a2a/a2a-service/requirements.txt
a2a-sdk>=0.3.0
uvicorn>=0.30.0
fastapi>=0.111.0
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.7.1
jinja2>=3.1.3
google-adk>=1.7.0
mcp>=1.5.0
asyncio>=3.4.3
asyncpg>=0.30.0
starlette>=0.37.2
web3>=6.15.1
eth-account>=0.9.0
rich>=13.7.0
typing-extensions>=4.12.2
sqlalchemy[asyncio]>=2.0.29
asyncpg>=0.29.0
greenlet>=3.0.3
openai>=2.0.0
crypto_com_facilitator_client>=1.0.3